requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2
dnspython==2.1.0
spacy==3.2.1
pyenchant==3.2.2
pickle4==0.0.1

